+++ src/components/App.css
@@ -1,0 +1,7 @@
+body, html, #root, .App {
+    background-color: #fafafa; 
+    width: 100vw;
+    font-family: var(--font-montserrat);
+    width: 100%;
+    margin: 0;
+}
