+++ src/containers/BookForm.js
@@ -2,8 +2,8 @@ import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
 import { addBook } from '../actions';
+import AddBook from '../components/AddBook';
 
-const categories = ['Action', 'Biography', 'History', 'Horror', 'Kids', 'Learning', 'Sci-Fi'];
 const defaultState = {
   category: '',
   title: '',
@@ -70,79 +70,16 @@ class BookForm extends Component {
           <p>{error}</p>
           )
         }
-        <form onSubmit={handleSubmit}>
-          <div className="input">
-            <div className="info">
-              <input
-                required
-                id="title"
-                className="title"
-                value={title}
-                minLength="2"
-                placeholder="Book Title"
-                onChange={e => handleChange('title', e.target.value)}
-              />
-              <input
-                required
-                id="author"
-                type="text"
-                autoComplete="name"
-                className="author"
-                value={author}
-                placeholder="Author"
-                onChange={e => handleChange('author', e.target.value)}
-              />
-              <select
-                id="category"
-                className="category"
-                onChange={e => handleChange('category', e.target.value)}
-                value={category}
-              >
-                {categories.map(thisCategory => (
-                  <option
-                    key={Math.random()}
-                    value={thisCategory}
-                  >
-                    {thisCategory}
-                  </option>
-                ))}
-              </select>
-            </div>
-            <div className="summary">
-              <textarea
-                id="summary"
-                value={summary}
-                placeholder="Summary"
-                onChange={e => handleChange('summary', e.target.value)}
-              />
-              <input
-                required
-                id="pages"
-                className="pages"
-                value={pages}
-                min="1"
-                max="2000"
-                type="number"
-                placeholder="Total Pages"
-                onChange={e => handleChange('pages', e.target.value)}
-              />
-              <input
-                required
-                id="progress"
-                className="progress"
-                min="0"
-                max="2000"
-                type="number"
-                value={progress}
-                placeholder="Current Page"
-                onChange={e => handleChange('progress', e.target.value)}
-              />
-            </div>
-          </div>
-          <div className="add-book">
-            <button type="submit">ADD BOOK</button>
-          </div>
-        </form>
+        <AddBook
+          category={category}
+          author={author}
+          title={title}
+          pages={pages}
+          progress={progress}
+          summary={summary}
+          handleChange={handleChange}
+          handleSubmit={handleSubmit}
+        />
       </div>
     );
   }
