+++ src/index.js
@@ -6,7 +6,6 @@ import App from './components/App';
 import * as serviceWorker from './serviceWorker';
 import rootReducer from './reducers';
 import './index.css';
-import './App.css';
 
 const defaultState = {
   filter: 'All',
